API Reference

GadgetIO.AbstractGadgetHeader

GadgetIO.GadgetCube

GadgetIO.GadgetCube

GadgetIO.GadgetCylinder

GadgetIO.GadgetSphere

GadgetIO.HaloID

GadgetIO.InfoLine

GadgetIO.KeyHeader

GadgetIO.SnapshotHeader

GadgetIO.SubfindHeader

GadgetIO.allocate_data_dict

GadgetIO.block_present

GadgetIO.check_block_position

GadgetIO.check_blocks

GadgetIO.check_blocksize

GadgetIO.check_in_cylinder

GadgetIO.check_info

GadgetIO.construct_matched_dict

GadgetIO.convert_string_to_int8

GadgetIO.filter_by_ids

GadgetIO.filter_cube

GadgetIO.filter_cylinder

GadgetIO.filter_ids

GadgetIO.filter_positions

GadgetIO.filter_sphere

GadgetIO.filter_subfind

GadgetIO.filter_subfind

GadgetIO.find_files_for_keys

GadgetIO.find_most_massive_halo

GadgetIO.find_read_positions

GadgetIO.find_read_positions

GadgetIO.find_read_positions

GadgetIO.get_block_positions

GadgetIO.get_datatype_name

GadgetIO.get_geometry_box_corners

GadgetIO.get_geometry_box_corners

GadgetIO.get_geometry_box_corners

GadgetIO.get_geometry_center

GadgetIO.get_geometry_center

GadgetIO.get_geometry_center

GadgetIO.get_geometry_mask

GadgetIO.get_geometry_mask

GadgetIO.get_geometry_mask

GadgetIO.get_index_bounds

GadgetIO.get_index_list

GadgetIO.get_index_list_arr

GadgetIO.get_index_list_dict

GadgetIO.get_index_list_set

GadgetIO.get_int8_blockname

GadgetIO.get_int_pos

GadgetIO.get_keylist

GadgetIO.get_lazy_vcat_indices

GadgetIO.get_pos_block_name

GadgetIO.get_prop_from_block

GadgetIO.get_rad_block_name

GadgetIO.get_requested_info

GadgetIO.get_total_particles

GadgetIO.head_to_dict

GadgetIO.head_to_obj

GadgetIO.join_blocks

GadgetIO.load_read_positions

GadgetIO.peano_hilbert_key

GadgetIO.print_blocks

GadgetIO.read_block

GadgetIO.read_block_data

GadgetIO.read_block_subsnaps

GadgetIO.read_block_with_offset!

GadgetIO.read_blocks_over_all_files

GadgetIO.read_bockname

GadgetIO.read_filtered

GadgetIO.read_halo_prop

GadgetIO.read_halo_prop

GadgetIO.read_halo_prop_and_id

GadgetIO.read_header

GadgetIO.read_ids_in_halo

GadgetIO.read_info

GadgetIO.read_info_line

GadgetIO.read_key_block

GadgetIO.read_key_index

GadgetIO.read_keyheader

GadgetIO.read_particles_by_id

GadgetIO.read_particles_by_id

GadgetIO.read_particles_by_id_single_file

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_box_peano

GadgetIO.read_particles_in_geometry

GadgetIO.read_particles_in_geometry

GadgetIO.read_particles_in_halo

GadgetIO.read_particles_in_halo

GadgetIO.read_particles_in_volume

GadgetIO.read_particles_in_volume

GadgetIO.read_pids

GadgetIO.read_snap

GadgetIO.read_subfind

GadgetIO.read_subfind

GadgetIO.read_subfind!

GadgetIO.read_subfind_header

GadgetIO.read_subsnaps

GadgetIO.reduce_read_positions

GadgetIO.save_read_positions

GadgetIO.select_file

GadgetIO.shift_across_box_border

GadgetIO.snap_to_dict

GadgetIO.write_block

GadgetIO.write_header

GadgetIO.write_info_block

Exported Functions

GadgetIO.block_present — Function

block_present(filename::String, blockname::String, blocks::Vector{String}=[""])

Checks if a given block is present in the snapshot file, or in the supplied blocks Vector.

GadgetIO.filter_cube — Method

filter_cube(snap_file::String, corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real}, 
                      parttype::Integer)

Reads positions from snap_file and returns the indices of particles contained in a box defined by corner_lowerleft and corner_upperright.

GadgetIO.filter_cylinder — Method

filter_cylinder(filename::String, pt1::Array{<:Real}, pt2::Array{<:Real}, r::Real)

Reads the positions contained in a file and returns the indices of particles contained in a cylinder defined by the endpoints pt1 and pt2 and radius r.

GadgetIO.filter_ids — Method

filter_ids(snap_file::String, selected_ids::Vector{<:Integer})

Reads IDs from snap_file and returns the indices of particles matching the selected_ids

GadgetIO.filter_subfind — Function

filter_subfind(filebase::String, blockname::String, filter_function::Function [, nfiles::Integer=1])

Selects entries in subfind block that fulfill the 'filter_funcion' requirements and returns a 'SubfindFilter' object.

Examples

julia> find_mass_gt_1e15(M) = ( (M > 1.e15) ? true : false )
find_mass_gt_1e15 (generic function with 1 method)
julia> filtered_subfind = filter_subfind(filebase, "MVIR", find_mass_gt_1e15)
[...]

GadgetIO.filter_subfind — Function

filter_subfind(filebase::String, filter_function::Function [, nfiles::Integer=1])

Selects entries in subfind block that fulfill the 'filter_funcion' requirements and returns a 'SubfindFilter' object.

Examples

julia> find_mass_gt_1e15(M) = ( (M > 1.e15) ? true : false )
find_mass_gt_1e15 (generic function with 1 method)
julia> filtered_subfind = filter_subfind(filebase, "MVIR", find_mass_gt_1e15)
[...]

GadgetIO.find_most_massive_halo — Function

find_most_massive_halo(filebase::String [, nfiles::Int=1])

Reads the selected file and its subfiles and returns position, virial radius and a HaloID object that contains the subfile which contains the most massive halo and the position in the block.

GadgetIO.find_read_positions — Method

find_read_positions( snap_base::String, geometry::AbstractGadgetGeometry;
                     parttype::Integer=0,
                     verbose::Bool=true )

Finds the number of particles and their storage location in a snapshot directory that are contained in the space defined by geometry.

GadgetIO.find_read_positions — Method

find_read_positions( snap_base::String, filter_function::Function)

Finds the number of particles and their storage location in a snapshot directory that pass the filter_function.

GadgetIO.find_read_positions — Method

find_read_positions(files::Vector{<:Integer}, filebase::String, 
                    blocks::AbstractVector{String}, parttype::Integer,
                    keylist::Vector{<:Integer}, key_info::Vector{InfoLine},
                    verbose::Bool)

Helper function to get positions and length of particle blocks in files.

GadgetIO.get_index_list — Method

get_index_list(list_to_find::Vector{<:Integer}, list_to_check::Vector{<:Integer})

Finds the indices at which list_to_check contains elements from list_to_find. If both either of the lists are not sorted it uses a Dict lookup, otherwise it uses a Vector forward-search.

GadgetIO.get_total_particles — Method

get_total_particles(h::AbstractGadgetHeader, parttype::Integer)

Calculates to total number of particles present in the simulation. Accounts for integer overflow.

GadgetIO.head_to_dict — Method

head_to_dict(filename::String)

Returns the header of a snapshot as a dictionary.

GadgetIO.head_to_obj — Method

head_to_obj(filename::String)

Returns the header of a snapshot as a SnapshotHeader object.

GadgetIO.load_read_positions — Method

load_read_positions(read_positions_file::String)

Loads the relevant read-in positions from a binary file.

GadgetIO.print_blocks — Method

print_blocks(filename::String; verbose::Bool=true)

Reads the block names of blocks in a snapshot and returns them in an array. Outputs them to console if verbose=true

GadgetIO.read_block — Method

read_block(filename::String, blockname::String;
                            info::InfoLine=InfoLine(),
                            parttype::Integer=-1)

Reads a block in a snapshot with given name. Names are case sensitive.

Examples

julia> pos_info = InfoLine("POS", Float32, 1, [1, 1, 1, 1, 1, 1])
[...]
julia> gas_pos = read_block(filename, "POS", info=pos_info, parttype=0)
[...]

GadgetIO.read_blocks_over_all_files — Method

read_blocks_over_all_files( snap_base::String, blocks::Array{String};
                            filter_function::Union{Function, Nothing}=nothing, 
                            read_positions::Union{Dict, Nothing}=nothing, 
                            parttype::Integer=0, verbose::Bool=true )

Reads the specified blocks from all distributed files where particles pass the filter_function, or are given by a Dict of read_positions. For read_positions please see find_read_positions.

GadgetIO.read_halo_prop — Method

read_halo_prop(filebase, haloid::HaloID, blockname::AbstractString; verbose::Bool=true)

Get halo property from block blockname by halo id. Returns an array or scalar depending on the block type.

GadgetIO.read_halo_prop — Method

read_halo_prop(filebase, i_global::Integer, blockname::AbstractString; verbose::Bool=true)

Get halo property from block blockname by global halo index i_global (zero-based index).

GadgetIO.read_halo_prop_and_id — Method

read_halo_prop_and_id(filebase, i_global::Integer, blockname::AbstractString; kwargs...)
read_halo_prop_and_id(filebase, i_global::Integer, blockname::AbstractString, nfiles::Integer; verbose::Bool=true)

Get halo property and HaloID from block blockname by global halo index i_global (zero-based index). nfiles should generally be set to h.num_files, obtained from read_header. When nfiles is not passed, it is read automatically from the header.

GadgetIO.read_header — Method

read_header(filename::String)

Reads the header of a snapshot file or file base (without .0, .1, etc.) and returns a SnapshotHeader object.

See also: head_to_obj

GadgetIO.read_ids_in_halo — Method

read_ids_in_halo( sub_base::String, halo::HaloID; 
                  halo_type::Integer=1, verbose::Bool=true)

Reads the IDs of all particles contained in a halo.

GadgetIO.read_info — Method

read_info(filename; verbose::Bool=false)

Reads the info block of a snapshot and returns the information in an array of InfoLine types. If verbose=true the blocknames are also printed to console.

GadgetIO.read_particles_by_id — Method

read_particles_by_id(snap_base::String, ids::Array{<:Integer}, 
                     blocks::Array{String}; 
                     parttype::Integer=0, verbose::Bool=true,
                     pos0::Array{<:Real}=[-1.234, -1.234, -1.234],
                     r0::Real=0.0)

Reads particles filtered by the provided IDs. Returns all requested blocks as entries in a Dict.

GadgetIO.read_particles_by_id — Method

read_particles_by_id(snap_base::String, ids::Array{<:Integer}, 
                     block::String; 
                     parttype::Integer=0, verbose::Bool=true,
                     pos0::Array{<:Real}=[-1.234, -1.234, -1.234],
                     r0::Real=0.0)

Reads particles filtered by the provided IDs. Returns the requested block as an Array.

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::String,
                      corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real};
                      parttype::Integer=0, verbose::Bool=true,
                      use_keys::Bool=true)

Like read_particles_in_box but for a single block. Returns the block as an array.

See also: read_particles_in_box

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::String,
                      corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real};
                      parttype::Integer=0, verbose::Bool=true,
                      use_keys::Bool=true)

Reads all particles within a box defined by a lower left and upper right corner for a given particle type. Returns a dictionary with all requested blocks. If use_keys=true it uses Peano-Hilbert keys to constrain the read-in, otherwise it uses a brute-force read-in with a filter function. Peano-Hilbert key based read-in is significantly faster.

GadgetIO.read_particles_in_geometry — Method

read_particles_in_geometry( filename::String, block::String,
                            geometry::AbstractGadgetGeometry;
                            parttype::Integer=0, verbose::Bool=true,
                            use_keys::Bool=true, do_shift_across_box_border::Bool=true)

Reads all particles within a space defined by an AbstractGeometry struct for a given particle type. Returns a dictionary with the requested block.

GadgetIO.read_particles_in_geometry — Method

read_particles_in_geometry( filename::String, blocks::Vector{String},
                            geometry::AbstractGadgetGeometry;
                            parttype::Integer=0, verbose::Bool=true,
                            use_keys::Bool=true, do_shift_across_box_border::Bool=true)

Reads all particles within a space defined by an AbstractGeometry struct for a given particle type. Returns a dictionary with all requested blocks. If shift_across_box_border is true, the particles are moved beyond the borders of a periodic box, if false, the periodicity of the box is still accounted for, but the particles' positions are not shifted.

GadgetIO.read_particles_in_halo — Method

read_particles_in_halo(snap_base::String, blocks::Array{String},
                            sub_base::String, halo::HaloID; 
                            radius::Union{Real,Nothing}=nothing,
                            rad_scale::Real=1.0, halo_type::Integer=1,
                            parttype::Integer=0, verbose::Bool=true,
                            use_keys::Bool=true)

Reads all particles of type parttype that are contained in a halo defined by its HaloID. If radius is given (in simulation units), particles are read within at least this radius times rad_scale. Otherwise, R200, RMEA, or RHMS times rad_scale is used depending on halo_type (1, 1, and 2, respectively). Returns a Dict with each of the blocks as entries.

GadgetIO.read_particles_in_halo — Method

read_particles_in_halo( snap_base::String, block::String,
                        sub_base::String, halo::HaloID; 
                        kwargs...)

Reads all particles of type parttype that are contained in a halo defined by its HaloID. If radius is given (in simulation units), particles are read within at least this radius times rad_scale. Otherwise, R200, RMEA, or RHMS times rad_scale is used depending on halo_type (1, 1, and 2, respectively). Returns an Array with the requested block.

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::String,
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true)

Like read_particles_in_volume but for a single block. Returns the block as an array.

See also: read_particles_in_volume

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::Vector{String},
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true,
                      use_keys::Bool=true)

Reads all particles within a box encapsulating a volume defined by center position and radius for a given particle type. Returns a dictionary with all requested blocks.

See also: read_particles_in_box

GadgetIO.read_snap — Function

read_snap(filename::String [, blockname::String="", parttype::Integer=-1] )

Wrapper function to read snapshot in various ways: filename only: returns the entire snapshot as a dictionary. blockname: Returns only that block. If parttype specified only for that particle type.

Examples

julia> gas_pos = read_snap(filename, "POS", 0)

GadgetIO.read_subfind — Method

read_subfind(filename::String, blockname::String, ids::AbstractVector{<:Integer}; return_haloid::Bool=false)

Reads the block at the given subfind indices (ids, 0-indexed).

If return_haloid is true, returns a tuple of the block array and the corresponding HaloIDs.

Example

# Read the virial masses of the first four halos in subfind:
mvir = read_subfind(filebase, "MVIR", [0, 1, 2, 3])

# or:
mvir, haloids = read_subfind(filebase, "MVIR", [0, 1, 2, 3]; return_haloid=true)

GadgetIO.read_subfind — Method

read_subfind(filename::String, blockname::String; return_haloid::Bool=false)

Reads a block of a subfind file.

If return_haloid is true, returns a tuple of the block array and the corresponding HaloIDs.

GadgetIO.read_subfind_header — Method

read_subfind_header(filename::String)

Reads the header of a subfind file or file base (without .0, .1, etc.) into a SubfindHeader struct.

GadgetIO.save_read_positions — Method

save_read_positions(read_positions_file::String, data)

Saves the relevant read-in positions to a binary file.

GadgetIO.snap_to_dict — Function

snap_to_dict(filename::String, try_info::Bool=true)

Reads whole snapshot into memory and returns a dictionary sorted by particle type and block names. Be cautious with large snapshots!

GadgetIO.write_block — Function

write_block(f::IOStream, data,
            blockname::String="";
            snap_format::Integer=2)

Write data to a block to an opened file f.

GadgetIO.write_header — Method

write_header(f::IOStream, h::SnapshotHeader; snap_format::Integer=2)

Writes the header block to an opened file f.

GadgetIO.write_info_block — Method

write_info_block(f::IOStream, info::Vector{InfoLine}; snap_format::Integer=2)

Writes the info block to an opened file f.

Exported Types

GadgetIO.AbstractGadgetHeader — Type

AbstractGadgetHeader

Abstract type to unify snapshot and subfind headers.

GadgetIO.GadgetCube — Type

struct GadgetCube{T} <: AbstractGadgetGeometry
    corner_lower_left::Vector{T}
    corner_upper_right::Vector{T}
end

Defines a cube by corner_lower_left and corner_upper_right. To be used for read_particles_in_geometry

GadgetIO.GadgetCube — Method

GadgetCube(center::Vector{T}, side_length::T) where T

Define a GadgetCube by its center and side length.

GadgetIO.GadgetCylinder — Type

struct GadgetCylinder{T} <: AbstractGadgetGeometry
    pos_start::Vector{T}
    pos_end::Vector{T}
    radius::T
end

Defines a cylinder by two end points pos_start and pos_start as well as a radius. To be used for read_particles_in_geometry

GadgetIO.GadgetSphere — Type

struct GadgetSphere{T} <: AbstractGadgetGeometry
    center::Vector{T}
    radius::T
end

Defines a sphere by center and radius. To be used for read_particles_in_geometry

GadgetIO.HaloID — Type

struct HaloID
    file::Int64
    id::Int64
end

Stores the subfile that contains the halo in file and the position in the block in id.

GadgetIO.InfoLine — Type

struct InfoLine([  block_name="", data_type=Float32, n_dim=Int32(0),
                            is_present=zeros(Int32, 6) ])

Contains the data of a single entry in the INFO block of a Gadget snapshot.

Fields

Name	Meaning

block_name::String	name of the data block, e.g. "POS"

data_type::DataType	datatype of the block, e.g. Float32 for single precision, Float64 for double

n_dim::Int32	number of dimensions of the block, usually 1 or 3

is_present::Vector{Int32}	array of flags for which particle type this block is present,

e.g. gas only: [ 1, 0, 0, 0, 0, 0 ],

or gas + BHs: [ 1, 0, 0, 0, 0, 1 ]

GadgetIO.SnapshotHeader — Type

mutable struct SnapshotHeader <: AbstractGadgetHeader

Contains the data of the HEAD block of a Gadget snapshot.

Fields

Name	Meaning

npart::Vector{Int32}	an array of particle numbers per type in this snapshot

massarr::Vector{Float64}	an array of particle masses per type in this snapshot - if zero: MASS block present

time::Float64	time / scale factor of the simulation

z::Float64	redshift of the simulation

flag_sfr::Int32	1 if simulation was run with star formation, else 0

flag_feedback::Int32	1 if simulation was run with stellar feedback, else 0

nall::Vector{UInt32}	total number of particles in the simulation

flag_cooling::Int32	1 if simulation was run with cooling, else 0

num_files::Int32	number of snapshots over which the simulation is distributed

omega_0::Float64	Omega matter

boxsize::Float64	total size of the simulation box

omega_l::Float64	Omega dark enery

h0::Float64	little h

flag_stellarage::Int32	1 if simulation was run with stellar age, else 0

flag_metals::Int32	1 if simulation was run with metals, else 0

npartTotalHighWord::Vector{UInt32}	If Npart > 1584^3 (>2^32) this contains a high bit: ntotal = npartTotalHighWord*2^32 + nall

flag_entropy_instead_u::Int32	1 if snapshot U field contains entropy instead of internal energy, else 0

flag_doubleprecision::Int32	1 if snapshot is in double precision, else 0

flag_ic_info::Int32	1 if initial snapshot file contains an info block, else 0

lpt_scalingfactor::Float32	factor to use second order ic generation

fill::Vector{Int32}	the HEAD block needs to be filled with zeros to have a size of 256 bytes

GadgetIO.SubfindHeader — Type

struct SubfindHeader

Contains the data of the HEAD block in the subfind output

Fields

Name	Meaning

nhalos::Int32	number of halos in the output file

nsubhalos::Int32	number of subhalos in the output file

nfof::Int32	number of particles in the FoF

ngroups::Int32	number of large groups in the output file

time::Float64	time / scale factor of the simulation

z::Float64	redshift of the simulation

tothalos::UInt32	total number of halos over all output files

totsubhalos::UInt32	total number of subhalos over all output files

totfof::UInt32	total number of particles in the FoF

totgroups::UInt32	1 if simulation was run with cooling, else 0

num_files::Int32	number of files over which subfind data is distributed

boxsize::Float64	total size of the simulation box

omega_0::Float64	Omega matter

omega_l::Float64	Omega dark enery

h0::Float64	little h

flag_doubleprecision::Int32	1 if snapshot is in double precision, else 0

flag_ic_info::Int32	1 if initial snapshot file contains an info block, else 0

Private Functions

GadgetIO.allocate_data_dict — Method

allocate_data_dict( blocks::Array{String}, N_to_read::Integer, 
                    snap_info::Array{InfoLine}, no_mass_block::Bool )

Helper function to allocate the data Dict.

GadgetIO.check_block_position — Method

check_block_position(filename::String, blockname::String)

Helper function to find read positions.

GadgetIO.check_blocks — Method

check_blocks(snap_base::String, blocks::Array{String})

Check if all requested blocks are present.

GadgetIO.check_blocksize — Method

check_blocksize(f::IOStream, position_before::Integer, blocksize_before::Integer)

Checks for integer overflow in the size of the block.

GadgetIO.check_in_cylinder — Method

check_in_cylinder(x, pt1, pt2, r)

Checks if a 3D point x is in a cylinder defined by its endpoints pt1 and pt2 and radius r.

GadgetIO.check_info — Method

check_info(filename::String, blockname::String)

Helper function to read INFO block or construct InfoLine for MASS block, if no INFO block is present. Returns a single InfoLine struct.

GadgetIO.construct_matched_dict — Method

construct_matched_dict(data_in::Dict{String, Union{Vector{T}, Array{T,2}}}, 
                            blocks::Array{String,1}, matched::Array{<:Integer,1}) where T

Write all matching particles to a new dictionary.

GadgetIO.convert_string_to_int8 — Method

convert_string_to_int8(name::String)

Converts a string to an array of Int8 to write to file.

GadgetIO.filter_by_ids — Method

filter_by_ids( selected_ids::Array{<:Integer}, ids::Array{<:Integer})

Checks for matching IDs.

GadgetIO.filter_positions — Method

filter_positions(snap_file::String, corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real}, 
                      parttype::Integer)

Reads positions from snap_file and returns the indices of particles contained in a box defined by corner_lowerleft and corner_upperright.

GadgetIO.filter_sphere — Method

filter_sphere(filename::String, center::Array{<:Real}, r::Real; parttype::Integer=0)

Reads the positions contained in a file and returns the indices of particles contained in a cylinder defined by the endpoints pt1 and pt2 and radius r.

GadgetIO.find_files_for_keys — Method

find_files_for_keys(filebase::String, nfiles::Integer, keylist::Vector{<:Integer})

Selects the files in which the particles associated with the given Peano-Hilbert keys are stored.

GadgetIO.get_block_positions — Method

get_block_positions(filename::String)

Returns a dictionary with the starting positions of all blocks in a snapshot in bits.

GadgetIO.get_datatype_name — Method

function get_datatype_name(dt::DataType, dim::Integer)

Convert the datatype into a string and then to a vector of 8bit integers to match name in Gadget.

GadgetIO.get_geometry_box_corners — Method

get_geometry_box_corners(cube::GadgetCube)

Returns a Tuple with the lower left and upper right corner of a box which contains the cube.

GadgetIO.get_geometry_box_corners — Method

get_geometry_box_corners(cylinder::GadgetCylinder)

Returns a tuple with the lower left and upper right corner of a box which contains the cylinder.

GadgetIO.get_geometry_box_corners — Method

get_geometry_box_corners(sphere::GadgetSphere)

Returns a tuple with the lower left and upper right corner of a box which contains the sphere.

GadgetIO.get_geometry_center — Method

function get_geometry_center(sphere::GadgetCube)

Returns the center of cube.

GadgetIO.get_geometry_center — Method

function get_geometry_center(cylinder::GadgetCylinder)

Returns the center of cylinder.

GadgetIO.get_geometry_center — Method

function get_geometry_center(sphere::GadgetSphere)

Returns the center of sphere.

GadgetIO.get_geometry_mask — Method

get_geometry_mask(cube::GadgetCube, pos::Matrix{T}) where T

Returns the indices of all particles contained in the cube.

GadgetIO.get_geometry_mask — Method

get_geometry_mask(cylinder::GadgetCylinder, pos::Matrix{T}) where T

Returns the indices of all particles contained in the cylinder.

GadgetIO.get_geometry_mask — Method

get_geometry_mask(sphere::GadgetSphere, pos::Matrix{T}) where T

Returns the indices of all particles contained in the sphere.

GadgetIO.get_index_bounds — Method

get_index_bounds(ids::Vector{<:Integer}, low_bounds::Vector{<:Integer}, high_bounds::Vector{<:Integer})

Returns sorted Vector of indices i, for which low_bounds[i] ≤ ids[j] ≤ high_bounds[i] for any j. All parameters ids, low_bounds, and high_bounds have to already be sorted.

GadgetIO.get_index_list_arr — Method

get_index_list_arr(list_to_find::Vector{<:Integer}, list_to_check::Vector{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses forward-searching in sorted array. Both arrays have to be sorted.

GadgetIO.get_index_list_dict — Method

get_index_list_dict(list_to_find::Vector{<:Integer}, list_to_check::Vector{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses a Dict for lookup -> slower than the array search, but works on unsorted arrays.

GadgetIO.get_index_list_set — Method

get_index_list_set(list_to_find::Vector{<:Integer}, list_to_check::Vector{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses a Set for lookup -> slower than the array search get_index_list_arr, but works on unsorted arrays like get_index_list_dict, just faster.

GadgetIO.get_int8_blockname — Method

get_int8_blockname(info::InfoLine)

Pad the stored block name with spaces and convert to Array of Int8.

GadgetIO.get_int_pos — Method

get_int_pos(pos::Real, domain_corner::Real, domain_fac::Real )

Computes the integer position along the PH line.

GadgetIO.get_keylist — Method

get_keylist(h_key::KeyHeader, x0::Array{<:Real}, x1::Array{<:Real})

Get all Peano-Hilbert keys for domain defined by the corner points x0 and x1.

GadgetIO.get_lazy_vcat_indices — Method

get_lazy_vcat_indices(arr::AbstractVector, inds)

For an array of arrays [a, b, c], where a, b, and c are arrays, returns the values of vcat(a, b, c)[ind].

This adapts to multi-dimensional arrays, respectively.

This is not exported.

GadgetIO.get_pos_block_name — Method

function get_pos_block_name(sub_file, halo_type)

Returns the position block name depending on the halo type.

GadgetIO.get_prop_from_block — Method

get_prop_from_block(block::Array, index::Integer)

Get the full property from block at index. Returns an array or scalar depending on the block type.

GadgetIO.get_rad_block_name — Method

function get_rad_block_name(sub_file, halo_type)

Returns the radius block name depending on the halo type.

GadgetIO.get_requested_info — Method

get_requested_info(snap_info::Array{InfoLine}, block::AbstractString)

Checks if the InfoLine is present for the requested block, or if the MASS block is requested, but not in the INFO block.

GadgetIO.join_blocks — Method

joinblocks(offsetkey, partperkey)

Joins neighboring blocks to simplify read-in.

GadgetIO.peano_hilbert_key — Method

peano_hilbert_key(bits::Integer, x::Integer, y::Integer, z::Integer)

Computes a Peano-Hilbert key for an integer triplet (x,y,z) with x,y,z typically in the range between 0 and 2^bits-1. Values outside this range can occur when reading across the periodic box borders.

GadgetIO.read_block_data — Method

read_block_data(f::IOStream, data_type::DataType, n_dim::Integer, npart::Integer)

Reads the binary data in a block.

GadgetIO.read_block_subsnaps — Method

read_block_subsnaps(filename::String, blockname::String;
                            info::InfoLine=InfoLine(),
                            parttype::Integer=-1)

Reads a block over all sub-snapshots. Names are case sensitive.

Examples

julia> pos_info = InfoLine("POS", Float32, 1, [1, 1, 1, 1, 1, 1])
[...]
julia> gas_pos = read_block(filename, "POS", info=pos_info, parttype=0)
[...]

GadgetIO.read_block_with_offset! — Method

read_block_with_offset!(data, n_read::Integer, filename::String, pos0::Integer, info::InfoLine,
                            offset::Integer, offset_key::Array{<:Integer}, 
                            part_per_key::Array{<:Integer} )

Read part of a block to pre-allocated array.

GadgetIO.read_bockname — Method

read_bockname(f::IOStream)

Reads the name of the Format 2 block.

GadgetIO.read_filtered — Method

read_filtered(snap_file, blockname, parttype, block_position, matched)

Helper function to return the correct filtered arrays.

GadgetIO.read_info_line — Method

read_info_line(f::IOStream)

Helper function to read the binary data into a InfoLine struct.

GadgetIO.read_key_block — Method

read_key_block( filename_keyfile::String, key_info::Vector{InfoLine}, 
                fields::Vector{String}, key::String, parttype::Integer)

Returns key block for different key. Valid values of key: KEY, NKEY, OKEY

GadgetIO.read_key_index — Method

read_key_index(file_key_index::String)

Reads the .key.index file.

GadgetIO.read_keyheader — Method

read_keyheader(filename::String)

Reads the header of a .key file.

GadgetIO.read_particles_by_id_single_file — Method

read_particles_by_id_single_file(snap_file::String, halo_ids::Array{<:Integer}, 
                                      blocks::Array{String}, parttype::Integer)

Helper function to read the matching IDs of particles in one whole file.

GadgetIO.read_particles_in_box_peano — Method

read_particles_in_box_peano(filename::String, blocks::Vector{String},
                            corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real};
                            parttype::Integer=0, verbose::Bool=true)

Reads all particles within a box defined by a lower left and upper right corner for a given particle type based on peano hilbert key reading. Returns a dictionary with all requested blocks.

GadgetIO.read_pids — Method

read_pids(sub_base::String, offset::Integer, N_ids::Integer)

Reads the PID block in the subfind output.

GadgetIO.read_subfind! — Method

read_subfind!(a::AbstractArray, filename, blockname::String)

Reads a block of a subfind file into a.

GadgetIO.read_subsnaps — Method

read_subsnaps(filebase::String, blockname::String, parttype::Integer,
                      info::InfoLine, h_global::SnapshotHeader)

Reads a block over distributed files and returns it in one large Array.

GadgetIO.reduce_read_positions — Method

reduce_read_positions(sel::Array{<:Integer})

Reduces the individual read positions by finding sub-ranges. Returns an array of read indices and an array with the number of entries per index.

GadgetIO.select_file — Method

select_file(filebase::String, filenum::Integer)

Checks if the requested files exists and returns the path.

GadgetIO.shift_across_box_border — Method

shift_across_box_border(x::Real, x_halo::Real, boxsize::Real, boxsize_half::Real)

Shift coordinate x across the box border if the zero coordinate x₀ is on the other side.

Private Types

GadgetIO.KeyHeader — Type

struct KeyHeader

Helper struct to store header information of .key files.

Fields

Name	Meaning

nkeys_file	Number of keys in this .key file

domain_corners	Corners of the domains defined by these keys

domain_fac	Factor needed for reconstructung int positions

bits	Size of PH keys it bits

nkeys_total	Total number of keys in all files

nkeys_total_highword	Total number of keys in all files