API Reference

GadgetIO.Header

GadgetIO.Info_Line

GadgetIO.block_present

GadgetIO.filter_subfind

GadgetIO.find_files_for_keys_old

GadgetIO.find_most_massive_halo

GadgetIO.get_block_positions

GadgetIO.get_int_pos

GadgetIO.head_to_dict

GadgetIO.head_to_obj

GadgetIO.print_blocks

GadgetIO.read_block_by_name

GadgetIO.read_header

GadgetIO.read_info

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_volume

GadgetIO.read_particles_in_volume

GadgetIO.read_snap

GadgetIO.snap_to_dict

GadgetIO.write_block

GadgetIO.write_header

Exported Functions

GadgetIO.block_present — Function

block_present(filename::String, blockname::String, blocks::Vector{String}=[""])

Checks if a given block is present in the snapshot file, or in the supplied blocks Vector.

GadgetIO.filter_subfind — Function

filter_subfind(filebase::String, blockname::String, filter_function::Function [, nfiles::Integer=1])

Selects entries in subfind block that fulfill the 'filter_funcion' requirements and returns a 'SubfindFilter' object.

Examples

julia> find_mass_gt_1e15(M) = ( (M > 1.e15) ? true : false )
find_mass_gt_1e15 (generic function with 1 method)
julia> filtered_subfind = filter_subfind(filebase, "MVIR", find_mass_gt_1e15)
[...]

GadgetIO.find_most_massive_halo — Function

find_most_massive_halo(filebase::String [, nfiles::Int=1])

Reads the selected file and its subfiles and returns position, virial radius and a SubfindID object that contains the subfile which contains the most massive halo and the position in the block.

GadgetIO.head_to_dict — Method

head_to_dict(filename::String)

Returns the header of a snapshot as a dictionary.

GadgetIO.head_to_obj — Method

head_to_obj(filename::String)

Returns the header of a snapshot as a Header object.

GadgetIO.print_blocks — Method

print_blocks(filename::String; verbose::Bool=true)

Reads the block names of blocks in a snapshot and returns them in an array. Outputs them to console if verbose=true

GadgetIO.read_block_by_name — Method

read_block_by_name(filename::String, blockname::String;
                            info::Info_Line=Info_Line(),
                            parttype::Integer=-1)

Reads a block in a snapshot with given name. Names are case sensitive.

Examples

julia> pos_info = Info_Line("POS", Float32, 1, [1, 1, 1, 1, 1, 1])
[...]
julia> gas_pos = read_block_by_name(filename, "POS", info=pos_info, parttype=0)
[...]

GadgetIO.read_header — Method

read_header(filename::String)

Reads the header of a snapshot and returns a Header object.

See also: head_to_obj

GadgetIO.read_info — Method

read_info(filename; verbose::Bool=false)

Reads the info block of a snapshot and returns the information in an array of Info_Line types. If verbose=true the blocknames are also printed to console.

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::Vector{String},
                      corner_lowerleft, corner_upperright;
                      parttype::Integer=0, verbose::Bool=true)

Reads all particles within a box defined by a lower left and upper right corner for a given particle type. Returns a dictionary with all requested blocks.

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::String,
                      corner_lowerleft, corner_upperright;
                      parttype::Integer=0, verbose::Bool=true)

Like read_particles_in_box but for a single block. Returns the block as an array.

See also: read_particles_in_box

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::Vector{String},
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true)

Reads all particles within a box encapsulating a volume defined by center position and radius for a given particle type. Returns a dictionary with all requested blocks.

See also: read_particles_in_box

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::String,
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true)

Like read_particles_in_volume but for a single block. Returns the block as an array.

See also: read_particles_in_volume

GadgetIO.read_snap — Function

read_snap(filename::String [, blockname::String="", parttype::Integer=-1] )

Wrapper function to read snapshot in various ways: filename only: returns the entire snapshot as a dictionary. blockname: Returns only that block. If parttype specified only for that particle type.

Examples

julia> gas_pos = read_block_by_name(filename, "POS", 0)
[...]

GadgetIO.snap_to_dict — Function

snap_to_dict(filename::String, try_info::Bool=true)

Reads whole snapshot into memory and returns a dictionary sorted by particle type and block names. Be cautious with large snapshots!

GadgetIO.write_block — Function

write_block(f::IOStream, data,
            blockname::String="";
            snap_format::Integer=2)

Write data to a block to an opened file f.

GadgetIO.write_header — Method

write_header(f::IOStream, h::Header; snap_format::Integer=2)

Writes the header block to an opened file f.

Exported Types

GadgetIO.Header — Type

mutable struct Header( [ **Fields ])

Contains the data of the HEAD block of a Gadget snapshot.

Fields

Name	Meaning

npart::Vector{Int32}	an array of particle numbers per type in this snapshot

massarr::Vector{Float64}	an array of particle masses per type in this snapshot - if zero: MASS block present

time::Float64	time / scale factor of the simulation

z::Float64	redshift of the simulation

flag_sfr::Int32	1 if simulation was run with star formation, else 0

flag_feedback::Int32	1 if simulation was run with stellar feedback, else 0

nall::Vector{UInt32}	total number of particles in the simulation

flag_cooling::Int32	1 if simulation was run with cooling, else 0

num_files::Int32	number of snapshots over which the simulation is distributed

omega_0::Float64	Omega matter

boxsize::Float64	total size of the simulation box

omega_l::Float64	Omega dark enery

h0::Float64	little h

flag_stellarage::Int32	1 if simulation was run with stellar age, else 0

flag_metals::Int32	1 if simulation was run with metals, else 0

npartTotalHighWord::Vector{UInt32}	weird

flag_entropy_instead_u::Int32	1 if snapshot U field contains entropy instead of internal energy, else 0

flag_doubleprecision::Int32	1 if snapshot is in double precision, else 0

flag_ic_info::Int32	1 if initial snapshot file contains an info block, else 0

lpt_scalingfactor::Float32	factor to use second order ic generation

fill::Vector{Int32}	the HEAD block needs to be filled with zeros to have a size of 256 bytes

GadgetIO.Info_Line — Type

mutable struct Info_Line([  block_name="", data_type=Float32, n_dim=Int32(0),
                            is_present=zeros(Int32, 6) ])

Contains the data of a single entry in the INFO block of a Gadget snapshot.

Fields

Name	Meaning

block_name::String	name of the data block, e.g. "POS"

data_type::DataType	datatype of the block, e.g. Float32 for single precision, Float64 for double

n_dim::Int32	number of dimensions of the block, usually 1 or 3

is_present::Vector{Int32}	array of flags for which particle type this block is present,

e.g. gas only: [ 1, 0, 0, 0, 0, 0 ],

or gas + BHs: [ 1, 0, 0, 0, 0, 1 ]

Private Functions

GadgetIO.find_files_for_keys_old — Method

This should be fixed! It's A LOT faster than the simpler version!

GadgetIO.get_block_positions — Method

get_block_positions(filename::String)

Returns a dictionary with the starting positions of all blocks in a snapshot in bits.

GadgetIO.get_int_pos — Method

Helper function for Peano Hilbert stuff

Private Types