API Reference

GadgetIO.HaloID

GadgetIO.InfoLine

GadgetIO.KeyHeader

GadgetIO.SnapshotHeader

GadgetIO.SubfindHeader

GadgetIO.allocate_data_dict

GadgetIO.block_present

GadgetIO.check_block_position

GadgetIO.check_blocks

GadgetIO.check_blocksize

GadgetIO.check_in_cylinder

GadgetIO.check_info

GadgetIO.construct_matched_dict

GadgetIO.filter_by_ids

GadgetIO.filter_cube

GadgetIO.filter_cylinder

GadgetIO.filter_positions

GadgetIO.filter_sphere

GadgetIO.filter_subfind

GadgetIO.filter_subfind

GadgetIO.find_files_for_keys

GadgetIO.find_files_for_keys_AR

GadgetIO.find_most_massive_halo

GadgetIO.find_read_positions

GadgetIO.find_read_positions

GadgetIO.get_block_positions

GadgetIO.get_index_list

GadgetIO.get_index_list_arr

GadgetIO.get_index_list_arr_sort

GadgetIO.get_index_list_dict

GadgetIO.get_index_list_set

GadgetIO.get_int_pos

GadgetIO.get_keylist

GadgetIO.get_prop_from_block

GadgetIO.head_to_dict

GadgetIO.head_to_obj

GadgetIO.join_blocks

GadgetIO.load_read_positions

GadgetIO.peano_hilbert_key

GadgetIO.print_blocks

GadgetIO.read_block

GadgetIO.read_block_data

GadgetIO.read_block_with_offset!

GadgetIO.read_blocks_over_all_files

GadgetIO.read_bockname

GadgetIO.read_filtered

GadgetIO.read_halo_prop

GadgetIO.read_halo_prop_and_id

GadgetIO.read_header

GadgetIO.read_ids_in_halo

GadgetIO.read_info

GadgetIO.read_info_line

GadgetIO.read_key_index

GadgetIO.read_keyheader

GadgetIO.read_particles_by_id

GadgetIO.read_particles_by_id

GadgetIO.read_particles_by_id_single_file

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_box_peano

GadgetIO.read_particles_in_halo

GadgetIO.read_particles_in_halo

GadgetIO.read_particles_in_volume

GadgetIO.read_particles_in_volume

GadgetIO.read_pids

GadgetIO.read_snap

GadgetIO.read_subfind

GadgetIO.read_subfind_header

GadgetIO.reduce_read_positions

GadgetIO.save_read_positions

GadgetIO.select_file

GadgetIO.snap_to_dict

GadgetIO.write_block

GadgetIO.write_header

Exported Functions

GadgetIO.block_present — Function

block_present(filename::String, blockname::String, blocks::Vector{String}=[""])

Checks if a given block is present in the snapshot file, or in the supplied blocks Vector.

GadgetIO.filter_cube — Method

filter_cube(snap_file::String, corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real}, 
                      parttype::Integer)

Reads positions from snap_file and returns the indices of particles contained in a box defined by corner_lowerleft and corner_upperright.

GadgetIO.filter_cylinder — Method

filter_cylinder(filename::String, pt1::Array{<:Real}, pt2::Array{<:Real}, r::Real)

Reads the positions contained in a file and returns the indices of particles contained in a cylinder defined by the endpoints pt1 and pt2 and radius r.

GadgetIO.filter_subfind — Function

filter_subfind(filebase::String, blockname::String, filter_function::Function [, nfiles::Integer=1])

Selects entries in subfind block that fulfill the 'filter_funcion' requirements and returns a 'SubfindFilter' object.

Examples

julia> find_mass_gt_1e15(M) = ( (M > 1.e15) ? true : false )
find_mass_gt_1e15 (generic function with 1 method)
julia> filtered_subfind = filter_subfind(filebase, "MVIR", find_mass_gt_1e15)
[...]

GadgetIO.filter_subfind — Function

filter_subfind(filebase::String, blockname::String, filter_function::Function [, nfiles::Integer=1])

Selects entries in subfind block that fulfill the 'filter_funcion' requirements and returns a 'SubfindFilter' object.

Examples

julia> find_mass_gt_1e15(M) = ( (M > 1.e15) ? true : false )
find_mass_gt_1e15 (generic function with 1 method)
julia> filtered_subfind = filter_subfind(filebase, "MVIR", find_mass_gt_1e15)
[...]

GadgetIO.find_most_massive_halo — Function

find_most_massive_halo(filebase::String [, nfiles::Int=1])

Reads the selected file and its subfiles and returns position, virial radius and a HaloID object that contains the subfile which contains the most massive halo and the position in the block.

GadgetIO.find_read_positions — Method

find_read_positions(files::Array{<:Integer}, filebase::String, 
                         blocks::Array{String}, parttype::Integer)

Helper function to get positions and length of particle blocks in files.

GadgetIO.find_read_positions — Method

find_read_positions( snap_base::String, filter_function::Function)

Finds the number of particles and their storage location in a snapshot directory that pass the filter_function.

GadgetIO.get_index_list — Method

get_index_list(list_to_find::Array{<:Integer}, list_to_check::Array{<:Integer})

Finds the indices at which list_to_check contains elements from list_to_find. If both either of the lists are not sorted it uses a Dict lookup, otherwise it uses a Array forward-search.

GadgetIO.head_to_dict — Method

head_to_dict(filename::String)

Returns the header of a snapshot as a dictionary.

GadgetIO.head_to_obj — Method

head_to_obj(filename::String)

Returns the header of a snapshot as a SnapshotHeader object.

GadgetIO.load_read_positions — Method

load_read_positions(read_positions_file::String)

Loads the relevant read-in positions from a binary file.

GadgetIO.print_blocks — Method

print_blocks(filename::String; verbose::Bool=true)

Reads the block names of blocks in a snapshot and returns them in an array. Outputs them to console if verbose=true

GadgetIO.read_block — Method

read_block(filename::String, blockname::String;
                            info::InfoLine=InfoLine(),
                            parttype::Integer=-1)

Reads a block in a snapshot with given name. Names are case sensitive.

Examples

julia> pos_info = InfoLine("POS", Float32, 1, [1, 1, 1, 1, 1, 1])
[...]
julia> gas_pos = read_block(filename, "POS", info=pos_info, parttype=0)
[...]

GadgetIO.read_blocks_over_all_files — Method

read_blocks_over_all_files( snap_base::String, blocks::Array{String};
                            filter_function::Union{Function, Nothing}=nothing, 
                            read_positions::Union{Dict, Nothing}=nothing, 
                            parttype::Integer=0, verbose::Bool=true )

Reads the specified blocks from all distributed files where particles pass the filter_function, or are given by a Dict of read_positions. For read_positions please see find_read_positions.

GadgetIO.read_halo_prop — Method

read_halo_prop(filebase::String, haloid::HaloID, blockname::String; verbose::Bool=true)

Get halo property from block blockname by halo id. Returns an array or scalar depending on the block type.

GadgetIO.read_halo_prop_and_id — Function

read_halo_prop_and_id(filebase::String, i_global::Integer, blockname::String, nfiles::Integer=1; verbose::Bool=true)

Get halo property from block blockname by global halo index i_global (zero-based index). nfiles should generally be set to h.num_files, obtained from read_header.

GadgetIO.read_header — Method

read_header(filename::String)

Reads the header of a snapshot and returns a SnapshotHeader object.

See also: head_to_obj

GadgetIO.read_ids_in_halo — Method

read_ids_in_halo( sub_base::String, halo::HaloID; 
                  halo_type::Integer=1, verbose::Bool=true)

Reads the IDs of all particles contained in a halo.

GadgetIO.read_info — Method

read_info(filename; verbose::Bool=false)

Reads the info block of a snapshot and returns the information in an array of InfoLine types. If verbose=true the blocknames are also printed to console.

GadgetIO.read_particles_by_id — Method

read_particles_by_id(snap_base::String, ids::Array{<:Integer}, 
                     blocks::Array{String}; 
                     parttype::Integer=0, verbose::Bool=true,
                     pos0::Array{<:Real}=[-1.234, -1.234, -1.234],
                     r0::Real=0.0)

Reads particles filtered by the provided IDs. Returns all requested blocks as entries in a Dict.

GadgetIO.read_particles_by_id — Method

read_particles_by_id(snap_base::String, ids::Array{<:Integer}, 
                     block::String; 
                     parttype::Integer=0, verbose::Bool=true,
                     pos0::Array{<:Real}=[-1.234, -1.234, -1.234],
                     r0::Real=0.0)

Reads particles filtered by the provided IDs. Returns the requested block as an Array.

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::String,
                      corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real};
                      parttype::Integer=0, verbose::Bool=true,
                      use_keys::Bool=true)

Like read_particles_in_box but for a single block. Returns the block as an array.

See also: read_particles_in_box

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::String,
                      corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real};
                      parttype::Integer=0, verbose::Bool=true,
                      use_keys::Bool=true)

Reads all particles within a box defined by a lower left and upper right corner for a given particle type. Returns a dictionary with all requested blocks. If use_keys=true it uses Peano-Hilbert keys to constrain the read-in, otherwise it uses a brute-force read-in with a filter function. Peano-Hilbert key based read-in is significantly faster.

GadgetIO.read_particles_in_halo — Method

read_particles_in_halo(snap_base::String, blocks::Array{String},
                            sub_base::String, halo::HaloID; 
                            rad_scale::Real=1.0, halo_type::Integer=1,
                            parttype::Integer=0, verbose::Bool=true)

Reads all particles of type parttype that are contained in a halo defined by its HaloID. Returns a Dict with each of the blocks as entries.

GadgetIO.read_particles_in_halo — Method

read_particles_in_halo( snap_base::String, block::String,
                        sub_base::String, halo::HaloID; 
                        rad_scale::Real=1.0, halo_type::Integer=1,
                        parttype::Integer=0, verbose::Bool=true )

Reads all particles of type parttype that are contained in a halo defined by its HaloID. Returns an Array with the requested block.

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::String,
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true)

Like read_particles_in_volume but for a single block. Returns the block as an array.

See also: read_particles_in_volume

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::Vector{String},
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true,
                      use_keys::Bool=true)

Reads all particles within a box encapsulating a volume defined by center position and radius for a given particle type. Returns a dictionary with all requested blocks.

See also: read_particles_in_box

GadgetIO.read_snap — Function

read_snap(filename::String [, blockname::String="", parttype::Integer=-1] )

Wrapper function to read snapshot in various ways: filename only: returns the entire snapshot as a dictionary. blockname: Returns only that block. If parttype specified only for that particle type.

Examples

julia> gas_pos = read_snap(filename, "POS", 0)

GadgetIO.read_subfind — Method

read_subfind(filename::String, blockname::String)

Reads a block of a subfind file.

GadgetIO.read_subfind_header — Method

read_subfind_header(filename::String)

Reads the header of a subfind file into a SubfindHeader struct.

GadgetIO.save_read_positions — Method

save_read_positions(read_positions_file::String, data)

Saves the relevant read-in positions to a binary file.

GadgetIO.snap_to_dict — Function

snap_to_dict(filename::String, try_info::Bool=true)

Reads whole snapshot into memory and returns a dictionary sorted by particle type and block names. Be cautious with large snapshots!

GadgetIO.write_block — Function

write_block(f::IOStream, data,
            blockname::String="";
            snap_format::Integer=2)

Write data to a block to an opened file f.

GadgetIO.write_header — Method

write_header(f::IOStream, h::SnapshotHeader; snap_format::Integer=2)

Writes the header block to an opened file f.

Exported Types

GadgetIO.HaloID — Type

struct HaloID
    file::Int64
    id::Int64
end

Stores the subfile that contains the halo in file and the position in the block in id.

GadgetIO.InfoLine — Type

struct InfoLine([  block_name="", data_type=Float32, n_dim=Int32(0),
                            is_present=zeros(Int32, 6) ])

Contains the data of a single entry in the INFO block of a Gadget snapshot.

Fields

Name	Meaning

block_name::String	name of the data block, e.g. "POS"

data_type::DataType	datatype of the block, e.g. Float32 for single precision, Float64 for double

n_dim::Int32	number of dimensions of the block, usually 1 or 3

is_present::Vector{Int32}	array of flags for which particle type this block is present,

e.g. gas only: [ 1, 0, 0, 0, 0, 0 ],

or gas + BHs: [ 1, 0, 0, 0, 0, 1 ]

GadgetIO.SnapshotHeader — Type

mutable struct SnapshotHeader( [ **Fields ])

Contains the data of the HEAD block of a Gadget snapshot.

Fields

Name	Meaning

npart::Vector{Int32}	an array of particle numbers per type in this snapshot

massarr::Vector{Float64}	an array of particle masses per type in this snapshot - if zero: MASS block present

time::Float64	time / scale factor of the simulation

z::Float64	redshift of the simulation

flag_sfr::Int32	1 if simulation was run with star formation, else 0

flag_feedback::Int32	1 if simulation was run with stellar feedback, else 0

nall::Vector{UInt32}	total number of particles in the simulation

flag_cooling::Int32	1 if simulation was run with cooling, else 0

num_files::Int32	number of snapshots over which the simulation is distributed

omega_0::Float64	Omega matter

boxsize::Float64	total size of the simulation box

omega_l::Float64	Omega dark enery

h0::Float64	little h

flag_stellarage::Int32	1 if simulation was run with stellar age, else 0

flag_metals::Int32	1 if simulation was run with metals, else 0

npartTotalHighWord::Vector{UInt32}	weird

flag_entropy_instead_u::Int32	1 if snapshot U field contains entropy instead of internal energy, else 0

flag_doubleprecision::Int32	1 if snapshot is in double precision, else 0

flag_ic_info::Int32	1 if initial snapshot file contains an info block, else 0

lpt_scalingfactor::Float32	factor to use second order ic generation

fill::Vector{Int32}	the HEAD block needs to be filled with zeros to have a size of 256 bytes

Private Functions

GadgetIO.allocate_data_dict — Method

allocate_data_dict( blocks::Array{String}, N_to_read::Integer, 
                    snap_info::Array{InfoLine}, no_mass_block::Bool )

Helper function to allocate the data Dict.

GadgetIO.check_block_position — Method

check_block_position(filename::String, blockname::String)

Helper function to find read positions.

GadgetIO.check_blocks — Method

check_blocks(snap_base::String, blocks::Array{String})

Check if all requested blocks are present.

GadgetIO.check_blocksize — Method

check_blocksize(f::IOStream, position_before::Integer, blocksize_before::Integer)

Checks for integer overflow in the size of the block.

GadgetIO.check_in_cylinder — Method

check_in_cylinder(x, pt1, pt2, r)

Checks if a 3D point x is in a cylinder defined by its endpoints pt1 and pt2 and radius r.

GadgetIO.check_info — Method

check_info(filename::String, blockname::String)

Helper function to read INFO block or construct InfoLine for MASS block, if no INFO block is present. Returns a single InfoLine struct.

GadgetIO.construct_matched_dict — Method

construct_matched_dict(data_in::Dict{String, Union{Vector{T}, Array{T,2}}}, 
                            blocks::Array{String,1}, matched::Array{<:Integer,1}) where T

Write all matching particles to a new dictionary.

GadgetIO.filter_by_ids — Method

filter_by_ids( selected_ids::Array{<:Integer}, ids::Array{<:Integer})

Checks for matching IDs.

GadgetIO.filter_positions — Method

filter_positions(snap_file::String, corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real}, 
                      parttype::Integer)

Reads positions from snap_file and returns the indices of particles contained in a box defined by corner_lowerleft and corner_upperright.

GadgetIO.filter_sphere — Method

filter_sphere(filename::String, center::Array{<:Real}, r::Real; parttype::Integer=0)

Reads the positions contained in a file and returns the indices of particles contained in a cylinder defined by the endpoints pt1 and pt2 and radius r.

GadgetIO.find_files_for_keys — Method

find_files_for_keys(filebase::String, nfiles::Integer, keylist::Vector{<:Integer})

Selects the files in which the particles associated with the given Peano-Hilbert keys are stored.

GadgetIO.find_files_for_keys_AR — Method

find_files_for_keys_AR(filebase::String, nfiles::Integer, keylist::Vector{<:Integer})

Selects the files in which the particles associated with the given Peano-Hilbert keys are stored. Version of Antonio. (Slower than Klaus' version!)

GadgetIO.get_block_positions — Method

get_block_positions(filename::String)

Returns a dictionary with the starting positions of all blocks in a snapshot in bits.

GadgetIO.get_index_list_arr — Method

get_index_list_arr(list_to_find::Array{<:Integer}, list_to_check::Array{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses forward-searching in sorted array.

GadgetIO.get_index_list_arr_sort — Method

get_index_list_arr(list_to_find::Array{<:Integer}, list_to_check::Array{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses forward-searching in sorted array.

GadgetIO.get_index_list_dict — Method

get_index_list_dict(list_to_find::Array{<:Integer}, list_to_check::Array{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses a Dict for lookup -> slower than the array search, but works on unsorted arrays.

GadgetIO.get_index_list_set — Method

get_index_list_set(list_to_find::Array{<:Integer}, list_to_check::Array{<:Integer})

Get positions in list_to_check where list_to_check matches list_to_find. Uses a Set for lookup -> slower than the array search, but works on unsorted arrays like Dict, just faster.

GadgetIO.get_int_pos — Method

get_int_pos(pos::Real, domain_corner::Real, domain_fac::Real )

Computes the integer position along the PH line.

GadgetIO.get_keylist — Method

get_keylist(h_key::KeyHeader, x0::Array{<:Real}, x1::Array{<:Real})

Get all Peano-Hilbert keys for domain defined by the corner points x0 and x1.

GadgetIO.get_prop_from_block — Method

get_prop_from_block(block::Array, index::Integer)

Get the full property from block at index. Returns an array or scalar depending on the block type.

GadgetIO.join_blocks — Method

function joinblocks(offsetkey, partperkey)

Joins neigboring blocks to simplify read-in.

GadgetIO.peano_hilbert_key — Method

peano_hilbert_key(bits::Integer, x::Integer, y::Integer, z::Integer)

Computes a Peano-Hilbert key for an integer triplet (x,y,z) with x,y,z in the range between 0 and 2^bits-1.

GadgetIO.read_block_data — Method

read_block_data(f::IOStream, data_type::DataType, n_dim::Integer, npart::Integer)

Reads the binary data in a block.

GadgetIO.read_block_with_offset! — Method

read_block_with_offset!(data, n_read::Integer, filename::String, pos0::Integer, info::InfoLine,
                            offset::Integer, offset_key::Array{<:Integer}, 
                            part_per_key::Array{<:Integer} )

Read part of a block to pre-allocated array.

GadgetIO.read_bockname — Method

read_bockname(f::IOStream)

Reads the name of the Format 2 block.

GadgetIO.read_filtered — Method

read_filtered(snap_file, blockname, parttype, block_position, matched)

Helper function to return the correct filtered arrays.

GadgetIO.read_info_line — Method

read_info_line(f::IOStream)

Helper function to read the binary data into a InfoLine struct.

GadgetIO.read_key_index — Method

read_key_index(file_key_index::String)

Reads the .key.index file.

GadgetIO.read_keyheader — Method

read_keyheader(filename::String)

Reads the header of a .key file.

GadgetIO.read_particles_by_id_single_file — Method

read_particles_by_id_single_file(snap_file::String, halo_ids::Array{<:Integer}, 
                                      blocks::Array{String}, parttype::Integer)

Helper function to read the matching IDs of particles in one whole file.

GadgetIO.read_particles_in_box_peano — Method

read_particles_in_box_peano(filename::String, blocks::Vector{String},
                            corner_lowerleft::Array{<:Real}, corner_upperright::Array{<:Real};
                            parttype::Integer=0, verbose::Bool=true)

Reads all particles within a box defined by a lower left and upper right corner for a given particle type based on peano hilbert key reading. Returns a dictionary with all requested blocks.

GadgetIO.read_pids — Method

read_pids(sub_base::String, offset::Integer, N_ids::Integer)

Reads the PID block in the subfind output.

GadgetIO.reduce_read_positions — Method

reduce_read_positions(sel::Array{<:Integer})

Reduces the individual read positions by finding sub-ranges. Returns an array of read indices and an array with the number of entries per index.

GadgetIO.select_file — Method

select_file(filebase::String, filenum::Integer)

Checks if the requested files exists and returns the path.

Private Types

GadgetIO.KeyHeader — Type

struct KeyHeader

Helper struct to store header information of .key files.

Fields

Name	Meaning

nkeys_file	Number of keys in this .key file

domain_corners	Corners of the domains defined by these keys

domain_fac	Factor needed for reconstructung int positions

bits	Size of PH keys it bits

nkeys_total	Total number of keys in all files

nkeys_total_highword	Total number of keys in all files

GadgetIO.SubfindHeader — Type

struct SubfindHeader

Contains the data of the HEAD block in the subfind output

Fields

Name	Meaning

nhalos::Int32	number of halos in the output file

nsubhalos::Int32	number of subhalos in the output file

nfof::Int32	number of particles in the FoF

ngroups::Int32	number of large groups in the output file

time::Float64	time / scale factor of the simulation

z::Float64	redshift of the simulation

tothalos::UInt32	total number of halos over all output files

totsubhalos::UInt32	total number of subhalos over all output files

totfof::UInt32	total number of particles in the FoF

totgroups::UInt32	1 if simulation was run with cooling, else 0

num_colors::Int32	number of colors

boxsize::Float64	total size of the simulation box

omega_0::Float64	Omega matter

omega_l::Float64	Omega dark enery

h0::Float64	little h

flag_doubleprecision::Int32	1 if snapshot is in double precision, else 0

flag_ic_info::Int32	1 if initial snapshot file contains an info block, else 0