API Reference

GadgetIO.HaloID

GadgetIO.Header

GadgetIO.Info_Line

GadgetIO.KeyHeader

GadgetIO.allocate_data_dict

GadgetIO.block_present

GadgetIO.check_blocks

GadgetIO.filter_by_ids

GadgetIO.filter_subfind

GadgetIO.find_files_for_keys

GadgetIO.find_files_for_keys_AR

GadgetIO.find_most_massive_halo

GadgetIO.find_read_positions

GadgetIO.get_block_positions

GadgetIO.get_index_list

GadgetIO.get_index_list_dict

GadgetIO.get_int_pos

GadgetIO.get_keylist

GadgetIO.head_to_dict

GadgetIO.head_to_obj

GadgetIO.join_blocks

GadgetIO.peano_hilbert_key

GadgetIO.print_blocks

GadgetIO.read_block_by_name

GadgetIO.read_block_with_offset!

GadgetIO.read_header

GadgetIO.read_ids_in_halo

GadgetIO.read_info

GadgetIO.read_key_index

GadgetIO.read_keyheader

GadgetIO.read_particles_by_id

GadgetIO.read_particles_by_id

GadgetIO.read_particles_by_id_single_file

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_box

GadgetIO.read_particles_in_halo

GadgetIO.read_particles_in_halo

GadgetIO.read_particles_in_volume

GadgetIO.read_particles_in_volume

GadgetIO.read_pids

GadgetIO.read_snap

GadgetIO.read_subfind

GadgetIO.read_subfind_header

GadgetIO.select_file

GadgetIO.snap_to_dict

GadgetIO.write_block

GadgetIO.write_header

Exported Functions

GadgetIO.block_present — Function

block_present(filename::String, blockname::String, blocks::Vector{String}=[""])

Checks if a given block is present in the snapshot file, or in the supplied blocks Vector.

GadgetIO.filter_subfind — Function

filter_subfind(filebase::String, blockname::String, filter_function::Function [, nfiles::Integer=1])

Selects entries in subfind block that fulfill the 'filter_funcion' requirements and returns a 'SubfindFilter' object.

Examples

julia> find_mass_gt_1e15(M) = ( (M > 1.e15) ? true : false )
find_mass_gt_1e15 (generic function with 1 method)
julia> filtered_subfind = filter_subfind(filebase, "MVIR", find_mass_gt_1e15)
[...]

GadgetIO.find_most_massive_halo — Function

find_most_massive_halo(filebase::String [, nfiles::Int=1])

Reads the selected file and its subfiles and returns position, virial radius and a HaloID object that contains the subfile which contains the most massive halo and the position in the block.

GadgetIO.get_index_list — Method

get_index_list(idarr1::Array{<:Integer}, idarr2::Array{<:Integer})

Get positions in idarr2 where idarr2 matches idarr1.

GadgetIO.head_to_dict — Method

head_to_dict(filename::String)

Returns the header of a snapshot as a dictionary.

GadgetIO.head_to_obj — Method

head_to_obj(filename::String)

Returns the header of a snapshot as a Header object.

GadgetIO.print_blocks — Method

print_blocks(filename::String; verbose::Bool=true)

Reads the block names of blocks in a snapshot and returns them in an array. Outputs them to console if verbose=true

GadgetIO.read_block_by_name — Method

read_block_by_name(filename::String, blockname::String;
                            info::Info_Line=Info_Line(),
                            parttype::Integer=-1)

Reads a block in a snapshot with given name. Names are case sensitive.

Examples

julia> pos_info = Info_Line("POS", Float32, 1, [1, 1, 1, 1, 1, 1])
[...]
julia> gas_pos = read_block_by_name(filename, "POS", info=pos_info, parttype=0)
[...]

GadgetIO.read_header — Method

read_header(filename::String)

Reads the header of a snapshot and returns a Header object.

See also: head_to_obj

GadgetIO.read_ids_in_halo — Method

read_ids_in_halo( sub_base::String, halo::HaloID; 
                  halo_type::Integer=1, verbose::Bool=true)

Reads the IDs of all particles contained in a halo.

GadgetIO.read_info — Method

read_info(filename; verbose::Bool=false)

Reads the info block of a snapshot and returns the information in an array of Info_Line types. If verbose=true the blocknames are also printed to console.

GadgetIO.read_particles_by_id — Method

read_particles_by_id(snap_base::String, ids::Array{<:Integer}, 
                     blocks::Array{String}; 
                     parttype::Integer=0, verbose::Bool=true,
                     pos0::Array{<:Real}=[-1.234, -1.234, -1.234],
                     r0::Real=0.0)

Reads particles filtered by the provided IDs. Returns all requested blocks as entries in a Dict.

GadgetIO.read_particles_by_id — Method

read_particles_by_id(snap_base::String, ids::Array{<:Integer}, 
                     block::String; 
                     parttype::Integer=0, verbose::Bool=true,
                     pos0::Array{<:Real}=[-1.234, -1.234, -1.234],
                     r0::Real=0.0)

Reads particles filtered by the provided IDs. Returns the requested block as an Array.

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::Vector{String},
                      corner_lowerleft, corner_upperright;
                      parttype::Integer=0, verbose::Bool=true)

Reads all particles within a box defined by a lower left and upper right corner for a given particle type. Returns a dictionary with all requested blocks.

GadgetIO.read_particles_in_box — Method

read_particles_in_box(filename::String, blocks::String,
                      corner_lowerleft, corner_upperright;
                      parttype::Integer=0, verbose::Bool=true)

Like read_particles_in_box but for a single block. Returns the block as an array.

See also: read_particles_in_box

GadgetIO.read_particles_in_halo — Method

read_particles_in_halo(snap_base::String, blocks::Array{String},
                            sub_base::String, halo::HaloID; 
                            rad_scale::Real=1.0, halo_type::Integer=1,
                            parttype::Integer=0, verbose::Bool=true)

Reads all particles of type parttype that are contained in a halo defined by its HaloID. Returns a Dict with each of the blocks as entries.

GadgetIO.read_particles_in_halo — Method

read_particles_in_halo( snap_base::String, block::String,
                        sub_base::String, halo::HaloID; 
                        rad_scale::Real=1.0, halo_type::Integer=1,
                        parttype::Integer=0, verbose::Bool=true )

Reads all particles of type parttype that are contained in a halo defined by its HaloID. Returns an Array with the requested block.

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::Vector{String},
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true)

Reads all particles within a box encapsulating a volume defined by center position and radius for a given particle type. Returns a dictionary with all requested blocks.

See also: read_particles_in_box

GadgetIO.read_particles_in_volume — Method

read_particles_in_box(filename::String, blocks::String,
                      center_pos, radius;
                      parttype::Integer=0, verbose::Bool=true)

Like read_particles_in_volume but for a single block. Returns the block as an array.

See also: read_particles_in_volume

GadgetIO.read_snap — Function

read_snap(filename::String [, blockname::String="", parttype::Integer=-1] )

Wrapper function to read snapshot in various ways: filename only: returns the entire snapshot as a dictionary. blockname: Returns only that block. If parttype specified only for that particle type.

Examples

julia> gas_pos = read_block_by_name(filename, "POS", 0)
[...]

GadgetIO.read_subfind — Method

read_subfind(filename::String, blockname::String)

Reads a block of a subfind file.

GadgetIO.read_subfind_header — Method

read_subfind_header(filename::String)

Reads the header of a subfind file into a SubfindHeader struct.

GadgetIO.snap_to_dict — Function

snap_to_dict(filename::String, try_info::Bool=true)

Reads whole snapshot into memory and returns a dictionary sorted by particle type and block names. Be cautious with large snapshots!

GadgetIO.write_block — Function

write_block(f::IOStream, data,
            blockname::String="";
            snap_format::Integer=2)

Write data to a block to an opened file f.

GadgetIO.write_header — Method

write_header(f::IOStream, h::Header; snap_format::Integer=2)

Writes the header block to an opened file f.

Exported Types

GadgetIO.HaloID — Type

struct HaloID
    file::Int64
    id::Int64
end

Stores the subfile that contains the halo in file and the position in the block in id.

GadgetIO.Header — Type

mutable struct Header( [ **Fields ])

Contains the data of the HEAD block of a Gadget snapshot.

Fields

Name	Meaning

npart::Vector{Int32}	an array of particle numbers per type in this snapshot

massarr::Vector{Float64}	an array of particle masses per type in this snapshot - if zero: MASS block present

time::Float64	time / scale factor of the simulation

z::Float64	redshift of the simulation

flag_sfr::Int32	1 if simulation was run with star formation, else 0

flag_feedback::Int32	1 if simulation was run with stellar feedback, else 0

nall::Vector{UInt32}	total number of particles in the simulation

flag_cooling::Int32	1 if simulation was run with cooling, else 0

num_files::Int32	number of snapshots over which the simulation is distributed

omega_0::Float64	Omega matter

boxsize::Float64	total size of the simulation box

omega_l::Float64	Omega dark enery

h0::Float64	little h

flag_stellarage::Int32	1 if simulation was run with stellar age, else 0

flag_metals::Int32	1 if simulation was run with metals, else 0

npartTotalHighWord::Vector{UInt32}	weird

flag_entropy_instead_u::Int32	1 if snapshot U field contains entropy instead of internal energy, else 0

flag_doubleprecision::Int32	1 if snapshot is in double precision, else 0

flag_ic_info::Int32	1 if initial snapshot file contains an info block, else 0

lpt_scalingfactor::Float32	factor to use second order ic generation

fill::Vector{Int32}	the HEAD block needs to be filled with zeros to have a size of 256 bytes

GadgetIO.Info_Line — Type

mutable struct Info_Line([  block_name="", data_type=Float32, n_dim=Int32(0),
                            is_present=zeros(Int32, 6) ])

Contains the data of a single entry in the INFO block of a Gadget snapshot.

Fields

Name	Meaning

block_name::String	name of the data block, e.g. "POS"

data_type::DataType	datatype of the block, e.g. Float32 for single precision, Float64 for double

n_dim::Int32	number of dimensions of the block, usually 1 or 3

is_present::Vector{Int32}	array of flags for which particle type this block is present,

e.g. gas only: [ 1, 0, 0, 0, 0, 0 ],

or gas + BHs: [ 1, 0, 0, 0, 0, 1 ]

Private Functions

GadgetIO.allocate_data_dict — Method

allocate_data_dict( blocks::Array{String}, N_to_read::Integer, 
                    snap_info::Array{Info_Line}, no_mass_block::Bool )

Helper function to allocate the data Dict.

GadgetIO.check_blocks — Method

check_blocks(snap_base::String, blocks::Array{String})

Check if all requested blocks are present.

GadgetIO.filter_by_ids — Method

filter_by_ids( selected_ids::Array{<:Integer}, ids::Array{<:Integer})

Checks for matching IDs.

GadgetIO.find_files_for_keys — Method

find_files_for_keys(filebase::String, nfiles::Integer, keylist::Vector{<:Integer})

Selects the files in which the particles associated with the given Peano-Hilbert keys are stored.

GadgetIO.find_files_for_keys_AR — Method

find_files_for_keys_AR(filebase::String, nfiles::Integer, keylist::Vector{<:Integer})

Selects the files in which the particles associated with the given Peano-Hilbert keys are stored. Version of Antonio. (Slower than Klaus' version!)

GadgetIO.find_read_positions — Method

find_read_positions(files::Array{<:Integer}, filebase::String, 
                         blocks::Array{String}, parttype::Integer)

Helper function to get positions and length of particle blocks in files.

GadgetIO.get_block_positions — Method

get_block_positions(filename::String)

Returns a dictionary with the starting positions of all blocks in a snapshot in bits.

GadgetIO.get_index_list_dict — Method

get_index_list_dict(keylist::Array{<:Integer}, keys_in_file::Array{<:Integer})

Get positions in keys_in_file where keys_in_file matches keylist. Uses a Dict for lookup -> slower than the normal version.

GadgetIO.get_int_pos — Method

get_int_pos(pos::Real, domain_corner::Real, domain_fac::Real )

Computes the integer position along the PH line.

GadgetIO.get_keylist — Method

get_keylist(h_key::KeyHeader, x0::Array{<:Real}, x1::Array{<:Real})

Get all Peano-Hilbert keys for domain defined by the corner points x0 and x1.

GadgetIO.join_blocks — Method

function joinblocks(offsetkey, partperkey)

Joins neigboring blocks to simplify read-in.

GadgetIO.peano_hilbert_key — Method

peano_hilbert_key(bits::Integer, x::Integer, y::Integer, z::Integer)

Computes a Peano-Hilbert key for an integer triplet (x,y,z) with x,y,z in the range between 0 and 2^bits-1.

GadgetIO.read_block_with_offset! — Method

read_block_with_offset!(data, n_read::Integer, filename::String, pos0::Integer, info::Info_Line,
                            offset::Integer, offset_key::Array{<:Integer}, 
                            part_per_key::Array{<:Integer} )

Read part of a block to pre-allocated array.

GadgetIO.read_key_index — Method

read_key_index(file_key_index::String)

Reads the .key.index file.

GadgetIO.read_keyheader — Method

read_keyheader(filename::String)

Reads the header of a .key file.

GadgetIO.read_particles_by_id_single_file — Method

read_particles_by_id_single_file(snap_file::String, halo_ids::Array{<:Integer}, 
                                      blocks::Array{String}, parttype::Integer)

Helper function to read the matching IDs of particles in one whole file.

GadgetIO.read_pids — Method

read_pids(sub_base::String, offset::Integer, N_ids::Integer)

Reads the PID block in the subfind output.

GadgetIO.select_file — Method

select_file(filebase::String, filenum::Integer)

Checks if the requested files exists and returns the path.

Private Types

GadgetIO.KeyHeader — Type

struct KeyHeader

Helper struct to store header information of .key files.

Fields

Name	Meaning

nkeys_file	Number of keys in this .key file

domain_corners	Corners of the domains defined by these keys

domain_fac	Factor needed for reconstructung int positions

bits	Size of PH keys it bits

nkeys_total	Total number of keys in all files

nkeys_total_highword	Total number of keys in all files